rename appInit to config store
