--- src/app/store/app-init/app-init.store.ts
@@ -1,7 +1,0 @@
-import { reducer } from './app-init.reducer';
-import { createFeature } from '@ngrx/store';
-
-export const fromAppInit = createFeature({
-  name: 'appInit',
-  reducer,
-});
