--- src/app/store/app-init/index.ts
@@ -1,2 +1,0 @@
-export * from './app-init.actions';
-export * as appInitEffects from './app-init.effects';
