+++ src/app/store/config/config.actions.ts
@@ -2,8 +2,8 @@ import { createActionGroup, emptyProps, props } from '@ngrx/store';
 import { AppConfig } from 'src/app/app.config';
 import { FullItemRoute } from 'src/app/models/routing/item-route';
 
-export const appInitActions = createActionGroup({
-  source: 'App init',
+export const configActions = createActionGroup({
+  source: 'Config',
   events: {
     init: emptyProps(),
     loadConfig: props<{ config: AppConfig }>(),
