+++ src/app/store/config/config.effects.ts
@@ -2,17 +2,17 @@ import { inject } from '@angular/core';
 import { Actions, createEffect, ofType, ROOT_EFFECTS_INIT } from '@ngrx/effects';
 import { map, take } from 'rxjs/operators';
 import { DEFAULT_ACTIVITY_ROUTE, DEFAULT_SKILL_ROUTE } from 'src/app/models/routing/default-route-tokens';
-import { appInitActions } from './app-init.actions';
+import { configActions } from './config.actions';
 import { APPCONFIG } from 'src/app/app.config';
 
 // Effect to trigger initialization once on app startup
-export const triggerAppInitEffect$ = createEffect(
+export const triggerConfigInitEffect$ = createEffect(
   (
     actions$ = inject(Actions)
   ) => actions$.pipe(
     ofType(ROOT_EFFECTS_INIT),
     take(1), // Ensure this runs only once
-    map(() => appInitActions.init())
+    map(() => configActions.init())
   ),
   { functional: true }
 );
@@ -21,10 +21,10 @@ export const loadConfigEffect$ = createEffect(
   (
     actions$ = inject(Actions)
   ) => actions$.pipe(
-    ofType(appInitActions.init),
+    ofType(configActions.init),
     map(() => {
       const config = inject(APPCONFIG);
-      return appInitActions.loadConfig({ config });
+      return configActions.loadConfig({ config });
     })
   ),
   { functional: true }
@@ -35,11 +35,11 @@ export const loadDefaultRoutesEffect$ = createEffect(
   (
     actions$ = inject(Actions)
   ) => actions$.pipe(
-    ofType(appInitActions.init),
+    ofType(configActions.init),
     map(() => {
       const defaultActivityRoute = inject(DEFAULT_ACTIVITY_ROUTE);
       const defaultSkillRoute = inject(DEFAULT_SKILL_ROUTE);
-      return appInitActions.setDefaultRoutes({ defaultActivityRoute, defaultSkillRoute });
+      return configActions.setDefaultRoutes({ defaultActivityRoute, defaultSkillRoute });
     })
   ),
   { functional: true }
