+++ src/app/store/config/config.reducer.ts
@@ -1,12 +1,12 @@
 import { createReducer, on } from '@ngrx/store';
-import { appInitActions } from './app-init.actions'; // Adjusted path
-import { initialState, State } from './app-init.state';
+import { configActions } from './config.actions';
+import { initialState, State } from './config.state';
 
 export const reducer = createReducer(
   initialState,
 
   on(
-    appInitActions.loadConfig,
+    configActions.loadConfig,
     (state, { config }): State => ({
       ...state,
       redirects: config.redirects,
