+++ src/app/store/config/config.store.ts
@@ -1,0 +1,7 @@
+import { createFeature } from '@ngrx/store';
+import { reducer } from './config.reducer';
+
+export const fromConfig = createFeature({
+  name: 'config',
+  reducer,
+});
