+++ src/app/store/config/index.ts
@@ -1,0 +1,3 @@
+export { fromConfig } from './config.store';
+export * as configEffects from './config.effects';
+export { configActions } from './config.actions';
